for #3814, remove all unnecessary type conversions of RemoteSysCpu and RemoteUserCpu and like attributes ===VersionHistory:Pending===

+++ src/condor_gridmanager/condorjob.cpp
@@ -1458,11 +1458,11 @@ ClassAd *CondorJob::buildSubmitAd()
 	submit_ad->Assign( ATTR_Q_DATE, now );
 	submit_ad->Assign( ATTR_CURRENT_HOSTS, 0 );
 	submit_ad->Assign( ATTR_COMPLETION_DATE, 0 );
-	submit_ad->Assign( ATTR_JOB_REMOTE_WALL_CLOCK, (float)0.0 );
-	submit_ad->Assign( ATTR_JOB_LOCAL_USER_CPU, (float)0.0 );
-	submit_ad->Assign( ATTR_JOB_LOCAL_SYS_CPU, (float)0.0 );
-	submit_ad->Assign( ATTR_JOB_REMOTE_USER_CPU, (float)0.0 );
-	submit_ad->Assign( ATTR_JOB_REMOTE_SYS_CPU, (float)0.0 );
+	submit_ad->Assign( ATTR_JOB_REMOTE_WALL_CLOCK, 0.0 );
+	submit_ad->Assign( ATTR_JOB_LOCAL_USER_CPU, 0.0 );
+	submit_ad->Assign( ATTR_JOB_LOCAL_SYS_CPU, 0.0 );
+	submit_ad->Assign( ATTR_JOB_REMOTE_USER_CPU, 0.0 );
+	submit_ad->Assign( ATTR_JOB_REMOTE_SYS_CPU, 0.0 );
 	submit_ad->Assign( ATTR_JOB_EXIT_STATUS, 0 );
 	submit_ad->Assign( ATTR_NUM_CKPTS, 0 );
 	submit_ad->Assign( ATTR_NUM_RESTARTS, 0 );
