+++ src/condor_status.V6/prettyPrint.cpp
@@ -71,8 +71,8 @@ static const char *formatDueDate( int , AttrList* , Formatter &);
 static const char *formatElapsedTime( int , AttrList* , Formatter &);
 static const char *formatRealTime( int , AttrList * , Formatter &);
 static const char *formatRealDate( int , AttrList * , Formatter &);
-//static const char *formatFloat (float, AttrList *, Formatter &);
-static const char *formatLoadAvg (float, AttrList *, Formatter &);
+//static const char *formatFloat (double, AttrList *, Formatter &);
+static const char *formatLoadAvg (double, AttrList *, Formatter &);
 
 static void ppInit()
 {
@@ -1293,7 +1293,7 @@ printCustom (ClassAd *ad)
 }
 
 static const char *
-formatLoadAvg (float fl, AttrList *, Formatter &)
+formatLoadAvg (double fl, AttrList *, Formatter &)
 {
 	static char buf[60];
 	sprintf(buf, "%.3f", fl);
